backend/requirements.txt
django
djangorestframework
Pillow
django-cors-headers
psycopg2